fix(export): remove artificial output limits and widen deterministic frame indexing
- Output FPS 合法性收敛为"正整数"：删除 Maximum=1000 与 RangeText，GUI 文案同步；
  Output FPS 是 output sampling density，性能 / 磁盘 / 预计时长都不构成合法性条件
- Application.targetFrameRate 改为 long 派生 + int.MaxValue 饱和，明确其为
  best-effort derived hint，绝不反向限制 Output FPS
- 新增 Export/SafetyFrameLimitPolicy：safety 默认 unbounded（不存在内建总帧数 /
  总时长上限），legacy 默认 36000 与 <= 0 统一解释为"未配置"，显式正值原样生效、
  不再夹取 1000000；判定统一走 IsFrameLimitReached
- canonical / output frame 编号与计数（MasterTimeline.FrameIndex / Prepare、
  _outputFrameIndex、capture request / captured / tail 计数、completionFrameIndex、
  resolvedTailFrames、FrameCaptureDriver 帧号与 PNG 文件名）从 int 收敛到 long，
  并加入结构性 fail-closed：帧号到达 long.MaxValue 时 frame-index-exhausted，
  End Tail 的 double→long 转换在 2^63 边界检查，均不依赖 unchecked 溢出
- metadata 记录实际采用的 policy（safetyPolicy / safetyFrameLimit /
  safetyDurationSeconds，unbounded 时为 null）；GUI 安全上限显示同步
- controller.paused fail-closed 从启动前移到 playback readiness 阶段：
  编辑器 idle 下 paused == true 是正常状态（scrController.Awake 执行
  paused = ADOBase.isLevelEditor，只有 editor.Play() → scnGame.Play() 才清零），
  启动前检查必然误杀正常导出；改为生命周期到达 PlayerControl + playerAlive 后
  仍 paused（或不可读）才 fail-closed，Renderist 任何阶段都不写 paused
- 同步 PROJECT_UNDERSTANDING.md：Output FPS 语义、safety 语义与 legacy 歧义、
  long 帧号链与 double 精度边界、paused 生命周期（§10.5）、实机验证结果与后续需求

+++ src/ADOFAI.Renderist/Export/FrameCaptureDriver.cs
@@ -28,8 +28,12 @@ namespace ADOFAI.Renderist.Export
     /// </summary>
     internal static class FrameCaptureDriver
     {
-        /// <summary>捕获结果回调（在 Unity 主线程 WaitForEndOfFrame 之后调用）。</summary>
-        public delegate void CaptureResultCallback(long generation, int frameIndex, bool success, string filePath, string error);
+        /// <summary>
+        /// 捕获结果回调（在 Unity 主线程 WaitForEndOfFrame 之后调用）。
+        /// frameIndex 是 canonical output frame number，类型为 long（合法 Output FPS
+        /// 为任意正 int，帧号不能依赖 int）。
+        /// </summary>
+        public delegate void CaptureResultCallback(long generation, long frameIndex, bool success, string filePath, string error);
 
         /// <summary>本阶段 Render Source 标签；写入 session metadata。</summary>
         public const string CameraSourceLabel = "scrCamera-rendertexture";
@@ -255,7 +259,7 @@ namespace ADOFAI.Renderist.Export
         /// generation 与当前 active generation 不一致、或 Render Source 尚未激活时返回 false。
         /// 绝不回退到 Screen framebuffer。
         /// </summary>
-        public static bool RequestCapture(long generation, int frameIndex)
+        public static bool RequestCapture(long generation, long frameIndex)
         {
             if (generation != _activeGeneration) return false;
             if (!_sourceActive || _captureTarget == null) return false;
@@ -516,7 +520,7 @@ namespace ADOFAI.Renderist.Export
             private long _generation;
 
             private bool _pending;
-            private int _pendingIndex;
+            private long _pendingIndex;
             private Texture2D _texture;
             private bool _stopped;
 
@@ -538,7 +542,7 @@ namespace ADOFAI.Renderist.Export
                 }
             }
 
-            public void RequestCapture(int frameIndex)
+            public void RequestCapture(long frameIndex)
             {
                 if (_stopped || _generation != _activeGeneration) return;
                 if (!_pending)
@@ -572,7 +576,7 @@ namespace ADOFAI.Renderist.Export
                     if (_stopped || _generation != _activeGeneration) yield break;
                     if (!_pending) continue;
 
-                    int index = _pendingIndex;
+                    long index = _pendingIndex;
                     _pending = false;
                     if (index == 0) Log.Info("MasterTimeline Stage=Frame0 AFTER_EOF frameIndex=0");
                     _pendingIndex = -1;
@@ -580,7 +584,7 @@ namespace ADOFAI.Renderist.Export
                 }
             }
 
-            private void CaptureNow(int frameIndex)
+            private void CaptureNow(long frameIndex)
             {
                 if (_stopped || _generation != _activeGeneration) return;
 
@@ -658,7 +662,12 @@ namespace ADOFAI.Renderist.Export
                 _texture = new Texture2D(_captureWidth, _captureHeight, TextureFormat.RGB24, false);
             }
 
-            private string BuildFilePath(int frameIndex)
+            /// <summary>
+            /// frame_&lt;index&gt;.png。index 是 long：ZeroPadWidth 只是**最小**补零宽度，
+            /// 位数超过它时自然扩展（PadLeft 只补不截），因此不存在编号截断，
+            /// 也绝不把 index 转回 int。
+            /// </summary>
+            private string BuildFilePath(long frameIndex)
             {
                 string indexText = frameIndex
                     .ToString(CultureInfo.InvariantCulture)
